Use k8s initContainers to migrate and seed db.
Using initContainers as opposed to seeding/migrating the DB in the
main function. This gives us a good separation of concerns, and keeps
main() focused on configuration.

The docker container now creates a binary from the admin tooling. On startup,
the initContainer runs the admin package to migrate and seed the DB.